coverage for new function
