+++ services/ui-src/src/utils/index.ts
@@ -10,6 +10,7 @@ export * from "./auth/authLifecycle";
 export * from "./autosave/autosave";
 // forms
 export * from "./forms/forms";
+export * from "./forms/dynamicItemFields";
 export * from "./forms/ilosFields";
 export * from "./forms/naaarPlanCompliance";
 export * from "./forms/priorAuthorization";
